mysite/requirements.txt
asgiref==3.3.4
Django==3.2.3
django-autoslug==1.9.8
django-crispy-forms==1.11.2
django-recaptcha==2.0.6
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
